Move test files to parameters folder.
